# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Authentication",
        "description": "API de gestion des utilisateurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/role/create": {
            "post": {
                "tags": [
                    "Rôle"
                ],
                "summary": "Créer un rôle",
                "description": "Crée un nouveau rôle. Réservé aux administrateurs.",
                "operationId": "b9d9348ad6a6ef24a69e0a6dc04fb0c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "ROLE_MANAGER"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rôle créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rôle créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/role/{id}": {
            "delete": {
                "tags": [
                    "Rôle"
                ],
                "summary": "Supprimer un rôle",
                "description": "Supprime un rôle par son identifiant. Réservé aux administrateurs.",
                "operationId": "0515de88901b4928d965d7de467e18c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du rôle",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rôle supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rôle supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Rôle introuvable"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}